Consolidate duplicate fs.stat().catch(() => null) pattern into safeStatOrNull helper
