+++ src/cli/src/shared/index.ts
@@ -1,6 +1,6 @@
 export * from "./ensure-dir.js";
 export * from "./error-guards.js";
-export * from "./fs-artifacts.js";
+export { safeStatOrNull, writeFileArtifact, writeJsonArtifact } from "./fs-artifacts.js";
 export * from "./module.js";
 export * from "./package-resolution.js";
 export * from "./repo-root.js";
